+++ src/lightly_train/_task_models/dinov2_ltdetr_object_detection/train_model.py
@@ -167,6 +167,7 @@ class DINOv2LTDETRObjectDetectionTrain(TrainModel):
             losses=model_args.criterion_losses,
             alpha=model_args.criterion_alpha,
             gamma=model_args.criterion_gamma,
+            num_classes=len(data_args.included_classes),
         )
 
         self.clip_max_norm = model_args.clip_max_norm