+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix `image_size` not tuple when training from pretrained model.
 - Fix a bug when fine-tuning a model with `resume_interrupted=True`.
+- Fix `num_classes` not updated when loading an object detection checkpoint with different number of classes.
 
 ### Security